refactor: move vscale client.
